Xbox and Page Refresh
Added URLs that point to Xbox. The page also refreshes every 5 seconds when it can't find the add to cart button. It's kind of slow for refreshing, but it's the fastest I could get it to go without being caught. The UI is GROSS, but it gets the job done. My code is also messy as I didn't rename variables in some places out of laziness, but it works. I plan on cleaning stuff up, but I wanted to push this code as I'm starting to get pull requests